api/requirements.txt
Flask==3.0.*
numpy==2.2.*
python-dotenv==1.1.*
google-genai==1.12.*
